# openapi.yaml
openapi: 3.0.0
info:
    title: Superdesk Publisher
    version: 2.x
paths:
...
    "/api/{version}/content/articles/{id}":
        get:
            summary: Show single tenant article
            parameters:
                - name: version
                  in: path
                  required: true
                  schema:
                      type: string
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      pattern: .+
            responses:
                "200":
                    description: Returned on success.
                    content:
                        "*/*":
                            schema:
                                $ref: "#/components/schemas/Article"
            tags:
                - article
        patch:
            summary: Updates articles
            description: >-
                Possible article statuses are:

                 * new
                 * published
                 * unpublished
                 * canceled

                Changing status from any status to `published` will make article visible for every user.


                Changing status from `published` to any other will make article hidden for user who don't have rights to see unpublished articles.
            parameters:
                - name: version
                  in: path
                  required: true
                  schema:
                      type: string
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      pattern: .+
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ArticleType"
                required: true
            responses:
                "200":
                    description: Returned on success.
                    content:
                        "*/*":
                            schema:
                                $ref: "#/components/schemas/Article"
                "400":
                    description: Returned when validation failed.
                "500":
                    description: Returned when unexpected error.
            tags:
                - article
        delete:
            summary: Deletes articles
            parameters:
                - name: version
                  in: path
                  required: true
                  schema:
                      type: string
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      pattern: .+
            responses:
                "204":
                    description: Returned on success.
                "404":
                    description: Returned when article not found.
                "500":
                    description: Returned when unexpected error.
            tags:
                - article
...
    "/api/{version}/content/push":
        post:
            summary: Adds a new content from HTTP Push
            parameters:
                - name: version
                  in: path
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                description: NinJS body
                required: true
            responses:
                "201":
                    description: Returned on success
            tags:
                - content push
    "/api/{version}/assets/push":
        post:
            summary: Adds new assets from HTTP Push
            parameters:
                - name: version
                  in: path
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/x-www-form-urlencoded:
                        schema:
                            type: object
                            properties:
                                media_id:
                                    type: string
                                media:
                                    type: string
                                    format: binary
            responses:
                "200":
                    description: Returned on form errors
                "201":
                    description: Returned on successful post.
                "500":
                    description: Returned on invalid file.
            tags:
                - content push
    "/api/{version}/assets/{action}/{mediaId}.{extension}":
        get:
            summary: Gets a single media file
            description: 'Action route parameter values: "get" or "push"'
            parameters:
                - name: version
                  in: path
                  required: true
                  schema:
                      type: string
                - name: action
                  in: path
                  required: true
                  schema:
                      type: string
                      pattern: get|push
                - name: mediaId
                  in: path
                  required: true
                  schema:
                      type: string
                      pattern: .+
                - name: extension
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Returned on form errors
                "404":
                    description: Returned when file doesn't exist.
            tags:
                - content push
...
    "/api/{version}/content/articles":
        patch:
            summary: Update article comments number
            parameters:
                - name: version
                  in: path
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ArticleCommentsType"
            responses:
                "200":
                    description: Returned on success.
                    content:
                        "*/*":
                            schema:
                                $ref: "#/components/schemas/Article"
                "404":
                    description: Return when article was not found
            tags:
                - article
...
    "/api/{version}/content/articles/":
        get:
            summary: Search articles
            parameters:
                - name: status
                  in: query
                  required: false
                  description: Package status
                  schema:
                      type: array
                      items:
                          type: string
                - name: published_before
                  in: query
                  required: false
                  description: The datetime before which the article has been published
                  schema:
                      type: string
                - name: published_after
                  in: query
                  required: false
                  description: The datetime after which the article has been published
                  schema:
                      type: string
                - name: author
                  in: query
                  required: false
                  description: Article authors
                  schema:
                      type: array
                      items:
                          type: string
                - name: keywords
                  in: query
                  required: false
                  description: Article keywords
                  schema:
                      type: array
                      items:
                          type: string
                - name: term
                  in: query
                  required: false
                  description: Search phrase
                  schema:
                      type: string
                - name: sorting
                  in: query
                  required: false
                  description: List order
                  schema:
                      type: array
                      items:
                          type: string
                - name: source
                  in: query
                  required: false
                  description: Sources
                  schema:
                      type: array
                      items:
                          type: string
                - name: limit
                  in: query
                  required: false
                  description: Items per page
                  schema:
                      type: integer
                - name: page
                  in: query
                  required: false
                  description: Page number
                  schema:
                      type: integer
                - name: route
                  in: query
                  required: false
                  description: Routes ids
                  schema:
                      type: array
                      items:
                          type: integer
                - name: metadata
                  in: query
                  required: false
                  description: "Metadata (e.g. query param:
            ?metadata[located][]=Sydney&metadata[located][]=Berlin)"
                  schema:
                      type: array
                      items:
                          type: string
                - name: version
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Returned on success.
                "500":
                    description: Unexpected error.
            tags:
                - article
...
security:
    - Bearer: []
servers:
    - url: http://www.publisher.wip
    - url: https://www.publisher.wip
components:
...
    securitySchemes:
        Bearer:
            name: Authorization
            type: apiKey
            in: header
            description: "Value: Bearer {jwt}"
    schemas:
...
        Article:
            properties:
                id:
                    type: integer
                title:
                    type: string
                body:
                    type: string
                slug:
                    type: string
                published_at:
                    type: string
                    format: date-time
                status:
                    type: string
                route:
                    $ref: "#/components/schemas/Route"
                template_name:
                    type: string
                publish_start_date:
                    type: string
                    format: date-time
                publish_end_date:
                    type: string
                    format: date-time
                is_publishable:
                    type: boolean
                metadata:
                    additionalProperties: true
                    type: object
                lead:
                    type: string
                code:
                    type: string
                extra:
                    additionalProperties: true
                    type: object
                slideshows:
                    $ref: "#/components/schemas/ArrayCollection"
                created_at:
                    type: string
                    format: date-time
                updated_at:
                    type: string
                    format: date-time
                keywords:
                    $ref: "#/components/schemas/ArrayCollection"
                seo_metadata:
                    $ref: "#/components/schemas/ArticleSeoMetadata"
                feature_media:
                    $ref: "#/components/schemas/ArticleMedia"
                is_published_fbia:
                    type: boolean
                article_statistics:
                    $ref: "#/components/schemas/ArticleStatistics"
                external_article:
                    $ref: "#/components/schemas/ExternalArticle"
                comments_count:
                    type: integer
                is_published_to_apple_news:
                    type: boolean
                apple_news_article:
                    $ref: "#/components/schemas/AppleNewsArticle"
                tenant:
                    $ref: "#/components/schemas/TenantInterface"
                paywall_secured:
                    type: boolean
            type: object
        ArticleType:
            required:
                - status
                - route
                - template_name
                - seoMetadata
                - paywallSecured
            properties:
                status:
                    type: string
                route:
                    type: string
                template_name:
                    type: string
                seoMetadata:
                    $ref: "#/components/schemas/SeoMetadataType"
                paywallSecured:
                    $ref: "#/components/schemas/BooleanType"
            type: object
        Route:
            properties:
                requirements:
                    additionalProperties: true
                    type: object
                id:
                    title: Unique id of this route.
                    type: integer
                static_prefix:
                    title: >-
                        Part of the URL that does not have parameters and thus can be used
                        to

                        naivly guess candidate routes.
                    description: Note that this field is not used by PHPCR-ODM
                    type: string
                variable_pattern:
                    title: Variable pattern part. The static part of the pattern is the id
                        without the prefix.
                    type: string
                parent:
                    type: integer
                lft:
                    type: integer
                rgt:
                    type: integer
                level:
                    type: integer
                template_name:
                    type: string
                articles_template_name:
                    type: string
                type:
                    type: string
                cache_time_in_seconds:
                    type: integer
                name:
                    type: string
                slug:
                    type: string
                position:
                    type: integer
                articles_count:
                    $ref: "#/components/schemas/ArticlesCountInterface"
                paywall_secured:
                    type: boolean
            type: object
...
        ArticleCommentsType:
            required:
                - commentsCount
            properties:
                url:
                    type: string
                id:
                    type: number
                commentsCount:
                    type: number
            type: object
...
        ArticleSeoMetadata:
            properties:
                meta_title:
                    type: string
                meta_description:
                    type: string
                og_title:
                    type: string
                og_description:
                    type: string
                twitter_title:
                    type: string
                twitter_description:
                    type: string
            type: object
...
        SeoMetadataType:
            required:
                - metaDescription
                - metaTitle
                - ogDescription
                - ogTitle
                - twitterDescription
                - twitterTitle
            properties:
                metaDescription:
                    type: string
                metaTitle:
                    type: string
                ogDescription:
                    type: string
                ogTitle:
                    type: string
                twitterDescription:
                    type: string
                twitterTitle:
                    type: string
            type: object
...
        ArrayCollection: []
        ArticleMedia:
            properties:
                id:
                    type: integer
                file:
                    $ref: "#/components/schemas/File"
                image:
                    $ref: "#/components/schemas/Image"
                description:
                    type: string
                by_line:
                    type: string
                alt_text:
                    type: string
                usage_terms:
                    type: string
                headline:
                    type: string
                copyright_holder:
                    type: string
                copyright_notice:
                    type: string
            type: object
        ArticleStatistics:
            properties:
                impressions_number:
                    type: integer
                page_views_number:
                    type: integer
                internal_click_rate:
                    type: number
                    format: float
                created_at:
                    type: string
                    format: date-time
                updated_at:
                    type: string
                    format: date-time
            type: object
        ExternalArticle:
            type: object
        AppleNewsArticle:
            type: object
        TenantInterface:
            type: object
        BooleanType:
            type: object
        ArticlesCountInterface:
            type: object
...
        File:
            properties:
                id:
                    type: string
                file_extension:
                    title: Uploaded file extension.
                    type: string
                asset_id:
                    type: string
            type: object
        Image:
            properties:
                id:
                    type: integer
                file_extension:
                    title: Uploaded file extension.
                    type: string
                asset_id:
                    type: string
                width:
                    type: integer
                height:
                    type: integer
                length:
                    type: number
                    format: float
            type: object
...